Fix lunduke hour link
When Bryan Lunduke moved his channel from youtube to odysee 2 years ago,
he set all of the videos in his youtube channel to private, thus
breaking the reference to it on the redox-os "talks" page.
This commit adds and utilizes an hugo shortcode to embed odysee links
into the redox website.